Minor English grammar fixes + use Is.gd shortener
Mostly corrected English plain text strings to use correct English syntax and grammar (I figure Nechigawara isn't a native speaker). Also changed the "just shorten URL" link to open a new tab with the shortened URL on is.gd; the use of their open API can be changed in the corresponding backend JS in order to display it on the information modal box instead of using this behavior.